requirements.txt
streamlit
matplotlib
seaborn
urlextract
wordcloud
pandas
nltk
collection


#--chatbot requirements--
# openai
google-generativeai>=0.7.2
python-dotenv